Modulize ipc functionality
