+++ swhks/src/main.rs
@@ -1,17 +1,11 @@
 use std::error::Error;
 use std::fs::Permissions;
-use std::hash::{DefaultHasher, Hash, Hasher};
-use std::io::Read;
-use std::os::unix::net::UnixListener;
 use std::{
     fs::{self},
-    io::Write,
     path::{Path, PathBuf},
-    process::Command,
 };
 
 use clap::Parser;
-use nix::unistd::daemon;
 use std::{
     env,
     os::unix::fs::PermissionsExt,
@@ -20,6 +14,8 @@ use std::{
 use sysinfo::System;
 use sysinfo::{ProcessExt, SystemExt};
 
+mod ipc;
+
 /// IPC Server for swhkd
 #[derive(Parser)]
 #[command(version, about, long_about = None)]
@@ -27,15 +23,6 @@ struct Args {
     /// Enable Debug Mode
     #[arg(short, long)]
     debug: bool,
-}
-
-/// Get the environment variables
-/// These would be requested from the default shell to make sure that the environment is up-to-date
-fn get_env() -> Result<String, Box<dyn std::error::Error>> {
-    let shell = std::env::var("SHELL")?;
-    let cmd = Command::new(shell).arg("-c").arg("env").output()?;
-    let stdout = String::from_utf8(cmd.stdout)?;
-    Ok(stdout)
 }
 
 fn main() -> std::io::Result<()> {
@@ -47,8 +34,6 @@ fn main() -> std::io::Result<()> {
         env_logger::Builder::from_env(env_logger::Env::default().default_filter_or("swhks=warn"))
             .init();
     }
-
-    let mut prev_hash = calculate_hash(String::new());
 
     let invoking_uid = get_uid().unwrap();
     let runtime_dir = format!("/run/user/{}", invoking_uid);
@@ -66,54 +51,9 @@ fn main() -> std::io::Result<()> {
         fs::remove_file(&sock_file_path)?;
     }
 
-    let listener = UnixListener::bind(sock_file_path)?;
-    let mut buff = [0; 1];
-    log::debug!("Listening for incoming connections...");
-
-    for stream in listener.incoming() {
-        match stream {
-            Ok(mut stream) => {
-                println!("Connection established!");
-                stream.read_exact(&mut buff)?;
-                println!("Received: {:?}", buff);
-                if buff == [1] {
-                    log::debug!("Received VERIFY message from swhkd");
-                    let _ = stream.write_all(prev_hash.to_string().as_bytes());
-                    log::debug!("Sent hash to swhkd");
-                    stream.flush()?;
-                    continue;
-                }
-                if buff == [2] {
-                    log::debug!("Received GET message from swhkd");
-                    let env = get_env().unwrap();
-                    if prev_hash == calculate_hash(env.clone()) {
-                        log::debug!("No changes in environment variables");
-                    } else {
-                        log::debug!("Changes in environment variables");
-                    }
-                    prev_hash = calculate_hash(env.clone());
-                    let _ = stream.write_all(env.as_bytes());
-                    stream.flush()?;
-                    continue;
-                }
-            }
-            Err(e) => {
-                log::error!("Error: {}", e);
-                break;
-            }
-        }
-    }
+    ipc::server_loop(&sock_file_path)?;
 
     Ok(())
-}
-
-/// Calculates a simple hash of the string
-/// Uses the DefaultHasher from the std::hash module which is not a cryptographically secure hash,
-/// however, it is good enough for our use case.
-pub fn calculate_hash(t: String) -> u64 {
-    let mut hasher = DefaultHasher::new();
-    t.hash(&mut hasher);
-    hasher.finish()
 }
 
 pub fn setup_swhks(invoking_uid: u32, runtime_path: PathBuf) {
